Refactor DB Client
- Move decorators to `decorators.py`
- Extract some queries
- Replace uses of `get_offset` method with function version.
