+++ db/models/mixins.py
@@ -1,3 +1,4 @@
+import datetime
 from typing import Optional
 
 from sqlalchemy import func, ForeignKey
@@ -37,7 +38,7 @@ class CountSubqueryMetadata:
 
 
 class CreatedAtMixin:
-    created_at: Mapped[timestamp] = mapped_column(
+    created_at: Mapped[datetime.datetime] = mapped_column(
         server_default=func.current_timestamp()
     )
