+++ db/queries/instantiations/util/refresh_all_materialized_views.py
@@ -1,0 +1,14 @@
+
+
+REFRESH_ALL_MATERIALIZED_VIEWS_QUERIES =             """
+        DO $$
+        DECLARE
+            rec RECORD;
+        BEGIN
+            FOR rec IN SELECT schemaname, matviewname FROM pg_matviews
+            LOOP
+                EXECUTE 'REFRESH MATERIALIZED VIEW ' || quote_ident(rec.schemaname) || '.' || quote_ident(rec.matviewname);
+            END LOOP;
+        END;
+        $$;
+        """
